+++ config/params.php
@@ -2,7 +2,6 @@
 
 declare(strict_types=1);
 
-use Yiisoft\Router\Debug\DebugRoutesCommand;
 use Yiisoft\Router\Debug\RouterCollector;
 use Yiisoft\Router\Debug\UrlMatcherInterfaceProxy;
 use Yiisoft\Router\UrlMatcherInterface;
@@ -14,11 +13,6 @@ return [
         ],
         'trackedServices' => [
             UrlMatcherInterface::class => [UrlMatcherInterfaceProxy::class, RouterCollector::class],
-        ],
-    ],
-    'yiisoft/yii-console' => [
-        'commands' => [
-            DebugRoutesCommand::COMMAND_NAME => DebugRoutesCommand::class,
         ],
     ],
 ];