+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Enh #265: Adapt summary data in debug collector (@rustamwin)
 - New #249: Add option to ignore method failure handler to the 'Router' middleware (@olegbaturin)
 - New #249: Add custom response handlers for the method failure responses to the 'Router' middleware (@olegbaturin)
+- Bug #268: Remove debug console command from package configuration (@vjik)
 
 ## 4.0.0 February 25, 2025